microservice/requirements.txt
Flask==2.3.2
opencv-python>=4.9.0
Pillow>=10.1.0
Werkzeug==2.3.6
requests==2.31.0
numpy>=1.21.0
gunicorn>=21.2.0